test/json/flatten-deeply-nested-array.test.ts
import {describe, expect, test} from "vitest";
import {flat} from "@/json/flatten-deeply-nested-array";

describe('Test flat array', () => {
    test('Test case 1', () => {
        expect(flat([1, 2, 3, [4, 5, 6], [7, 8, [9, 10, 11], 12], [13, 14, 15]], 0))
            .toEqual([1, 2, 3, [4, 5, 6], [7, 8, [9, 10, 11], 12], [13, 14, 15]]);
    })
    test('Test case 2', () => {
        expect(flat([1, 2, 3, [4, 5, 6], [7, 8, [9, 10, 11], 12], [13, 14, 15]], 1))
            .toEqual([1, 2, 3, 4, 5, 6, 7, 8, [9, 10, 11], 12, 13, 14, 15])
    })
    test('Test case 3', () => {
        expect(flat([[1, 2, 3], [4, 5, 6], [7, 8, [9, 10, 11], 12], [13, 14, 15]], 2))
            .toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15])
    })
});